# Conflicts:
#   client/src/main/java/client/scenes/BoardOverviewController.java
#   client/src/main/java/client/scenes/CardOverviewController.java
#   client/src/main/java/client/scenes/CustomizationOverviewController.java
#   client/src/main/java/client/scenes/InitialOverviewController.java
#   client/src/main/java/client/scenes/TagController.java
#   client/src/main/resources/client/scenes/BoardOverview.fxml
#   client/src/main/resources/client/scenes/CardOverview.fxml
#   client/src/main/resources/client/scenes/HomePageOverview.fxml

+++ client/src/main/java/client/scenes/CardOverviewController.java
@@ -9,16 +9,14 @@ import jakarta.ws.rs.WebApplicationException;
 import javafx.application.Platform;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.Node;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyEvent;
-import javafx.scene.layout.Background;
-import javafx.scene.layout.BackgroundFill;
-import javafx.scene.layout.HBox;
-import javafx.scene.layout.VBox;
+import javafx.scene.layout.*;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Rectangle;
 import javafx.stage.Modality;
@@ -32,13 +30,8 @@ import java.util.List;
 import static com.google.inject.Guice.createInjector;
 
 public class CardOverviewController {
-    private Stage primaryStage;
-    private Scene overview;
     private static final Injector INJECTOR = createInjector(new MyModule());
     private static final MyFXML FXML = new MyFXML(INJECTOR);
-    private long cardId;
-    private ServerUtils server;
-
     public VBox vBox;
     public HBox hbox;
     public Listing list;
@@ -48,29 +41,74 @@ public class CardOverviewController {
     public TextArea description;
     @FXML
     public VBox colorSchemes;
+    @FXML
+    private AnchorPane anchorPane;
+    @FXML
+    private Button addDescriptionButton;
+    @FXML
+    private Button cardNameButton;
+    @FXML
+    private Button addSubtaskButton;
+    @FXML
+    private Button addTagButton;
+    @FXML
+    private Label subtaskLabel;
+    @FXML
+    private Label tagsLabel;
+    @FXML
+    private Label colorSchemesLabel;
 
+    private boolean hasAccess = true;
+    private Stage primaryStage;
+    private Scene overview;
+    private long cardId;
+    private Card curCard;
+    private ServerUtils server;
+    private boolean isRunning = false;
     private Board board = new Board("test", "", "");
     private String fileName = "user_files/temp.txt";
 
+    /**
+     * Setter for the server.
+     *
+     * @param server - the server to assign this card to
+     */
+    @Inject
+    public CardOverviewController(ServerUtils server) {
+        this.server = server;
+    }
+
+    public CardOverviewController() {}
+
     public void initialize() {
         server.registerForUpdatesSubtask(subTask -> Platform.runLater(this::refresh));
-        server.registerForUpdatesTag(tag -> Platform.runLater(this::refresh));
-        server.registerForUpdatesCard(card -> Platform.runLater(() -> {
-            boolean check = server.checkCard(card);
-            System.out.println("First: " + check);
-
-            if (check) {
-                System.out.println(check);
-                Alert alert = new Alert(Alert.AlertType.WARNING);
-                alert.setTitle("Card has been deleted.");
-                alert.setContentText("We are very sorry, but the card you are currently editing has been deleted by another user");
-                alert.showAndWait();
-                switchBoard((Stage) cardLabel.getScene().getWindow());
-                server.stop();
-            } else {
+        isRunning = true;
+        server.registerForMessages("/topic/colors", ColorScheme.class, q -> Platform.runLater(() -> {
+            if (isRunning) {
+                System.out.println("I am here");
                 refresh();
             }
         }));
+//        server.registerForUpdatesTag(tag -> {
+//            Platform.runLater(this::refresh);
+//        });
+        server.registerForUpdatesCard(card -> {
+            if (isRunning) {
+                Platform.runLater(() -> {
+                    if (cardId != card.getCardId()) {
+                        board = server.getBoardByID(board.getBoardId());
+                        List<Listing> lists = board.getLists();
+                        for (Listing l : lists) {
+                            for (Card c : l.getCards()) {
+                                if (c.getName().equals(curCard.getName())) {
+                                    setList(l);
+                                    setCardId(c.getCardId());
+                                }
+                            }
+                        }
+                    }
+                    refresh();
+                });}});
     }
 
     /**
@@ -82,7 +120,6 @@ public class CardOverviewController {
         this.list = list;
     }
 
-
     /**
      * Setter for the board.
      *
@@ -90,6 +127,11 @@ public class CardOverviewController {
      */
     public void setBoard(Board board) {
         this.board = board;
+    }
+
+    public void setHasAccess(boolean check)
+    {
+        hasAccess = check;
     }
 
     /**
@@ -102,26 +144,13 @@ public class CardOverviewController {
     }
 
     /**
-     * Setter for the server.
-     *
-     * @param server - the server to assign this card to
-     */
-    @Inject
-    public CardOverviewController(ServerUtils server) {
-        this.server = server;
-    }
-
-    public CardOverviewController() {
-
-    }
-
-    /**
      * Setter for the current card.
      *
      * @param cardId the card's id
      */
     public void setCardId(long cardId) {
         this.cardId = cardId;
+        this.curCard = server.getCardsById(cardId);
     }
 
     /**
@@ -131,11 +160,14 @@ public class CardOverviewController {
      * @throws IOException the exception which might be caused
      */
 
-    public void switchToBoardScene(javafx.event.Event event) throws IOException {
-        switchBoard((Stage) ((Node) event.getSource()).getScene().getWindow());
+    public void switchToBoardScene(javafx.event.ActionEvent actionEvent) throws IOException {
+        isRunning = false;
+        switchBoard((Stage) ((Node) actionEvent.getSource()).getScene().getWindow());
     }
 
     public void switchBoard(Stage stage) {
+        if (stage == null)
+            return;
         var cardOverview = FXML.load(BoardOverviewController.class, "client", "scenes", "BoardOverview.fxml");
         cardOverview.getKey().setFileName(fileName);
         cardOverview.getKey().setBoard(board);
@@ -153,6 +185,13 @@ public class CardOverviewController {
      * @param actionEvent the event
      */
     public void addDescription(javafx.event.ActionEvent actionEvent) {
+        if (!hasAccess) {
+            Alert alert = new Alert(Alert.AlertType.WARNING);
+            alert.setTitle("No access!");
+            alert.setContentText("Cant edit in read-only mode!");
+            alert.showAndWait();
+            return;
+        }
         Card card = server.getCardsById(cardId);
         String text = description.getText();
         card.setDescription(text);
@@ -167,6 +206,13 @@ public class CardOverviewController {
      * @param actionEvent the event
      */
     public void updateName(javafx.event.ActionEvent actionEvent) {
+        if (!hasAccess) {
+            Alert alert = new Alert(Alert.AlertType.WARNING);
+            alert.setTitle("No access!");
+            alert.setContentText("Cant edit in read-only mode!");
+            alert.showAndWait();
+            return;
+        }
         Card card = server.getCardsById(cardId);
         TextInputDialog dialog = new TextInputDialog(card.getName());
         dialog.setTitle("Change the name of the card");
@@ -192,6 +238,13 @@ public class CardOverviewController {
      * Method for addSubTask button in Card Details scene.
      */
     public void addSubTask() {
+        if (!hasAccess) {
+            Alert alert = new Alert(Alert.AlertType.WARNING);
+            alert.setTitle("No access!");
+            alert.setContentText("Cant edit in read-only mode!");
+            alert.showAndWait();
+            return;
+        }
         Card card = server.getCardsById(cardId);
 
         TextInputDialog dialog = new TextInputDialog();
@@ -221,6 +274,13 @@ public class CardOverviewController {
      * @param subTask     the subtask to be edited
      */
     private void editSubTask(ActionEvent actionEvent, SubTask subTask) {
+        if (!hasAccess) {
+            Alert alert = new Alert(Alert.AlertType.WARNING);
+            alert.setTitle("No access!");
+            alert.setContentText("Cant edit in read-only mode!");
+            alert.showAndWait();
+            return;
+        }
         Card card = server.getCardsById(cardId);
         //   SubTask subTask = server.getSubtaskById(subtaskId);
         TextInputDialog dialog = new TextInputDialog();
@@ -308,7 +368,7 @@ public class CardOverviewController {
         HBox hBoxButtons = new HBox(editST, deleteST, up, down);
 
         HBox hBox = new HBox();
-        hBox.setSpacing(100);
+        hBox.setSpacing(10);
         hBox.getChildren().addAll(hBoxCB, hBoxButtons);
 
         vBox.getChildren().add(hBox);
@@ -322,19 +382,26 @@ public class CardOverviewController {
      * @param cardid  the card of the subtask
      */
     private void moveDown(ActionEvent event, SubTask subTask, long cardid) {
-        List<SubTask> subTasks = new ArrayList<>();
+        if (!hasAccess) {
+            Alert alert = new Alert(Alert.AlertType.WARNING);
+            alert.setTitle("No access!");
+            alert.setContentText("Cant edit in read-only mode!");
+            alert.showAndWait();
+            return;
+        }
+        List<SubTask> sbtask = new ArrayList<>();
         Card card = server.getCardsById(cardId);
         int idx = card.getSubTasks().indexOf(subTask);
         System.out.println(idx);
         if (idx < card.getSubTasks().size() - 1) {
             for (int i = 0; i < card.getSubTasks().size(); i++) {
-                subTasks.add(card.getSubTasks().get(i));
+                sbtask.add(card.getSubTasks().get(i));
                 server.deleteSubtask(card.getSubTasks().get(i));
             }
             vBox.getChildren().clear();
-            subTasks.set(idx, card.getSubTasks().get(idx + 1));
-            subTasks.set(idx + 1, subTask);
-            for (SubTask task : subTasks) {
+            sbtask.set(idx, card.getSubTasks().get(idx + 1));
+            sbtask.set(idx + 1, sbtask);
+            for (SubTask task : sbtask) {
                 try {
                     server.sendCard(card);
                     server.saveSubtask(task);
@@ -360,19 +427,26 @@ public class CardOverviewController {
      */
 
     private void moveUp(ActionEvent event, SubTask subTask, long cardid) {
-        List<SubTask> subTasks = new ArrayList<>();
+        if (!hasAccess) {
+            Alert alert = new Alert(Alert.AlertType.WARNING);
+            alert.setTitle("No access!");
+            alert.setContentText("Cant edit in read-only mode!");
+            alert.showAndWait();
+            return;
+        }
+        List<SubTask> sbtask = new ArrayList<>();
         Card card = server.getCardsById(cardId);
         int idx = card.getSubTasks().indexOf(subTask);
         System.out.println(idx);
         if (idx > 0) {
             for (int i = 0; i < card.getSubTasks().size(); i++) {
-                subTasks.add(card.getSubTasks().get(i));
+                sbtask.add(card.getSubTasks().get(i));
                 server.deleteSubtask(card.getSubTasks().get(i));
             }
             vBox.getChildren().clear();
-            subTasks.set(idx, card.getSubTasks().get(idx - 1));
-            subTasks.set(idx - 1, subTask);
-            for (SubTask task : subTasks) {
+            sbtask.set(idx, card.getSubTasks().get(idx - 1));
+            sbtask.set(idx - 1, subTask);
+            for (SubTask task : sbtask) {
                 try {
                     server.sendCard(card);
                     server.saveSubtask(task);
@@ -383,12 +457,9 @@ public class CardOverviewController {
                     alert.showAndWait();
                 }
             }
-//             server.sendList(this.list);
-//            server.saveCard(card);
             refresh();
         }
     }
-    //}
 
     /**
      * Deleting a subtask from database.
@@ -398,6 +469,13 @@ public class CardOverviewController {
      */
 
     private void deleteSubTask(ActionEvent actionEvent, SubTask subtask) {
+        if (!hasAccess) {
+            Alert alert = new Alert(Alert.AlertType.WARNING);
+            alert.setTitle("No access!");
+            alert.setContentText("Cant edit in read-only mode!");
+            alert.showAndWait();
+            return;
+        }
         HBox clicked = (HBox) ((Button) actionEvent.getSource()).getParent();
         HBox subtsk = (HBox) clicked.getParent();
         VBox vbox = (VBox) subtsk.getParent();
@@ -419,16 +497,46 @@ public class CardOverviewController {
      * Method that refreshes all card components.
      */
     public void refresh() {
+        // setting the background color of the card details the same as the card scheme
+
+        Color background = Color.web(curCard.getBackgroundColor());
+        anchorPane.setBackground(new Background(new BackgroundFill(background, new CornerRadii(0), new Insets(0) )));
+
+        Color fontColor = Color.web(curCard.getFontColor());
+        cardLabel.setTextFill(fontColor);
+        subtaskLabel.setTextFill(fontColor);
+        tagsLabel.setTextFill(fontColor);
+        colorSchemesLabel.setTextFill(fontColor);
+
+
+        System.out.println("Refreshed refresh: " + 0);
+        curCard = server.getCardsById(cardId);
+        if (curCard == null) {
+            isRunning = false;
+            Alert alert = new Alert(Alert.AlertType.WARNING);
+            alert.setTitle("Card has been deleted.");
+            alert.setContentText("We are very sorry, but the card you are currently editing has been deleted by another user");
+            alert.showAndWait();
+            switchBoard((Stage) cardLabel.getScene().getWindow());
+            server.stop();
+            return;
+        }
+
+
+        board = server.getBoardByID(board.getBoardId());
         refreshSubTasks();
         refreshCardDetails();
         refreshTags();
         refreshSchemes();
+
+        setUpButtons();
     }
+
 
     public void refreshSchemes() {
         colorSchemes.getChildren().clear();
         Card card = server.getCardsById(cardId);
-        List<ColorScheme> schemes = list.getBoard().getSchemes();
+        List<ColorScheme> schemes = board.getSchemes();
         for (ColorScheme s : schemes) {
             HBox hBox = new HBox(10);
             hBox.setMinSize(150, 20);
@@ -466,11 +574,20 @@ public class CardOverviewController {
         apply.setOnMouseEntered(event -> apply.setStyle("-fx-background-color: green; -fx-text-fill: white; -fx-font-size: 8 px"));
         apply.setOnMouseExited(event -> apply.setStyle("-fx-background-color: white; -fx-text-fill: green; -fx-font-size: 8 px"));
         apply.setOnAction(event -> {
+            if (!hasAccess)
+            {
+                Alert alert = new Alert(Alert.AlertType.WARNING);
+                alert.setTitle("No access!");
+                alert.setContentText("Cant edit in read-only mode!");
+                alert.showAndWait();
+                return;
+            }
             Card card = server.getCardsById(cardId);
             card.setBackgroundColor(scheme.getBackgroundColor());
             card.setFontColor(scheme.getFontColor());
+            card.setSchemeName(scheme.getName());
             server.sendList(list);
-            server.saveCard(card);
+            server.saveCard(card, false);
             refresh();
         });
     }
@@ -482,8 +599,10 @@ public class CardOverviewController {
         for (Tag tag : card.getTags()) {
             Label tagLabel = new Label(tag.getTitle());
             Color color = Color.web(tag.getColor());
-            Background background = new Background(new BackgroundFill(color, null, null));
+            Color bg = color.deriveColor(1, 1, 1, 0.5);
+            Background background = new Background(new BackgroundFill(bg, new CornerRadii(5), null));
             tagLabel.setBackground(background);
+            tagLabel.setBorder(new Border(new BorderStroke(color, BorderStrokeStyle.SOLID, new CornerRadii(5), new BorderWidths(3))));
             tagLabel.setAlignment(Pos.CENTER);
             //  tagLabel.setStyle("-fx-background-radius: 20;");
             tagLabel.setMinSize(100, 40);
@@ -524,6 +643,14 @@ public class CardOverviewController {
      * @throws IOException possible error
      */
     public void switchToChooseTagScene(javafx.event.ActionEvent actionEvent) throws IOException {
+        if (!hasAccess)
+        {
+            Alert alert = new Alert(Alert.AlertType.WARNING);
+            alert.setTitle("No access!");
+            alert.setContentText("Cant edit in read-only mode!");
+            alert.showAndWait();
+            return;
+        }
         var chooseTagOverview = FXML.load(ChooseTagController.class, "client", "scenes", "ChooseTag.fxml");
         chooseTagOverview.getKey().setFileName(fileName);
         chooseTagOverview.getKey().setBoard(board);
@@ -549,5 +676,35 @@ public class CardOverviewController {
                 throw new RuntimeException(e);
             }
         }
+    
+     * This method applies styling to buttons from the card overview.
+     */
+    public void setUpButtons(){
+        colorButton(addDescriptionButton);
+        colorButton(cardNameButton);
+        colorButton(addSubtaskButton);
+        colorButton(addTagButton);
+    }
+
+    /**
+     * Colors a certain button in the card overview.
+     *
+     * @param button the button to color
+     */
+    private void colorButton(Button button) {
+        String backgroundColor = curCard.getBackgroundColor();
+        String fontColor = curCard.getFontColor();
+
+        button.setStyle("-fx-background-color:" + backgroundColor);
+        button.setTextFill(Color.web(fontColor));
+        button.setBorder(new Border(new BorderStroke(Color.web(fontColor), BorderStrokeStyle.SOLID, new CornerRadii(5), BorderWidths.DEFAULT)));
+        button.setOnMouseEntered(event -> {
+            button.setStyle("-fx-background-color:" + fontColor);
+            button.setTextFill(Color.web(backgroundColor));
+        });
+        button.setOnMouseExited(event -> {
+            button.setStyle("-fx-background-color:" + backgroundColor);
+            button.setTextFill(Color.web(fontColor));
+        });
     }
 }
